demo/rag_demo/requirements.txt
streamlit==1.34.0
# Use stable version of llama-index
llama-index==0.9.48
# Dependencies
chromadb==0.4.22
sentence-transformers==2.5.0
pymupdf==1.23.25
python-docx==1.1.0
pandas==2.2.0
pillow==10.2.0
transformers==4.39.0
torch==2.2.0
numpy==1.26.4
